Manage the Thales HSM user
It was previously managed in the Kolla containers:
https://review.opendev.org/#/c/626034

But this wasn't ported to the barbican packaging distgit; and it should
be managed here; not in the Dockerfile. We already manage the Barbican
user; so this patch will create the nfast user; and add it to the
barbican group as well.

Change-Id: Ic091cb506c4e61b43024f846b5f13910f2289584